+++ executer.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"context"
 	"database/sql"
-	"fmt"
 	"io"
 	"strings"
 	"sync"
@@ -15,42 +14,7 @@ import (
 	"github.com/pkg/errors"
 )
 
-// Config is a connection setting to MySQL.
-// Exists to generate a Golang connection DSN to MySQL
-type Config struct {
-	DSN      string
-	User     string
-	Password string
-	Host     string
-	Port     int
-	Database string
-}
-
-// NewDefaultConfig returns the config for connecting to the local MySQL server
-func NewDefaultConfig() *Config {
-	return &Config{
-		User: "root",
-		Host: "127.0.0.1",
-		Port: 3306,
-	}
-}
-
-//GetDSN returns a DSN dedicated to connecting to MySQL.
-func (c *Config) GetDSN() string {
-	if c.DSN == "" {
-		return fmt.Sprintf(
-			"%s:%s@tcp(%s:%d)/%s?parseTime=true",
-			c.User,
-			c.Password,
-			c.Host,
-			c.Port,
-			c.Database,
-		)
-	}
-	return strings.TrimPrefix(c.DSN, "mysql://")
-}
-
-//Executer queries the DB. There is no parallelism
+// Executer queries the DB. There is no parallelism
 type Executer struct {
 	mu              sync.Mutex
 	db              *sql.DB
@@ -61,12 +25,12 @@ type Executer struct {
 	timeCheckQuery  string
 }
 
-//New return Executer with config
+// New return Executer with config
 func New(config *Config) (*Executer, error) {
 	return Open(config.GetDSN())
 }
 
-//Open with dsn
+// Open with dsn
 func Open(dsn string) (*Executer, error) {
 	db, err := sql.Open("mysql", dsn)
 	if err != nil {
@@ -96,12 +60,12 @@ func (e *Executer) Close() error {
 	return e.db.Close()
 }
 
-//Execute SQL
+// Execute SQL
 func (e *Executer) Execute(queryReader io.Reader) error {
 	return e.ExecuteContext(context.Background(), queryReader)
 }
 
-//ExecuteContext SQL execute with context.Context
+// ExecuteContext SQL execute with context.Context
 func (e *Executer) ExecuteContext(ctx context.Context, queryReader io.Reader) error {
 	e.mu.Lock()
 	defer e.mu.Unlock()
@@ -200,32 +164,32 @@ func (e *Executer) queryContext(ctx context.Context, query string) error {
 	return nil
 }
 
-//LastExecuteTime returns last execute time on DB
+// LastExecuteTime returns last execute time on DB
 func (e *Executer) LastExecuteTime() time.Time {
 	return e.lastExecuteTime
 }
 
-//SetExecuteHook set non select query hook
+// SetExecuteHook set non select query hook
 func (e *Executer) SetExecuteHook(hook func(query string, rowsAffected, lastInsertId int64)) {
 	e.executeHook = hook
 }
 
-//SetSelectHook set select query hook
+// SetSelectHook set select query hook
 func (e *Executer) SetSelectHook(hook func(query string, columns []string, rows [][]string)) {
 	e.selectHook = hook
 }
 
-//SetIsSelectFunc :Set the function to decide whether to execute in QueryContext
+// SetIsSelectFunc :Set the function to decide whether to execute in QueryContext
 func (e *Executer) SetIsSelectFunc(f func(query string) bool) {
 	e.isSelectFunc = f
 }
 
-//SetTimeCheckQuery set time check query for non mysql db
+// SetTimeCheckQuery set time check query for non mysql db
 func (e *Executer) SetTimeCheckQuery(query string) {
 	e.timeCheckQuery = query
 }
 
-//SetTimeCheckQuery set select query hook, but result is table string
+// SetTimeCheckQuery set select query hook, but result is table string
 func (e *Executer) SetTableSelectHook(hook func(query, table string)) {
 	e.selectHook = func(query string, columns []string, rows [][]string) {
 		var buf strings.Builder
@@ -242,7 +206,7 @@ type QueryScanner struct {
 	*bufio.Scanner
 }
 
-//NewQueryScanner returns QueryScanner
+// NewQueryScanner returns QueryScanner
 func NewQueryScanner(queryReader io.Reader) *QueryScanner {
 	scanner := bufio.NewScanner(queryReader)
 	onSplit := func(data []byte, atEOF bool) (advance int, token []byte, err error) {
@@ -265,7 +229,7 @@ func NewQueryScanner(queryReader io.Reader) *QueryScanner {
 	}
 }
 
-//Query return
+// Query return
 func (s *QueryScanner) Query() string {
 	return strings.Trim(strings.NewReplacer(
 		"\r\n", " ",
